Factor out blockbookFetch utility from sendTransaction
