+++ src/endpoints/sendTransaction.ts
@@ -1,18 +1,14 @@
 import { asObject, asString } from 'cleaners'
-import fetch from 'node-fetch'
 import parse from 'url-parse'
 
 import { config } from '../config'
 import { JsonRpc, JsonRpcResponse, WrapperIo } from '../types'
+import { blockbookFetch } from '../util/blockbookFetch'
 
+const server = config.blockbookServer
 const asSendTransactionParams = asObject({
   hex: asString
 })
-
-// const asSendTransactionResponse = asObject({
-//   result: asString
-// })
-const server = config.blockbookServer
 
 export type GetTransactionParams = ReturnType<typeof asSendTransactionParams>
 
@@ -22,41 +18,13 @@ export const sendTransaction = async (
 ): Promise<JsonRpcResponse> => {
   const params = asSendTransactionParams(data.params)
   const { hex } = params
-
   const parsed = parse(server, true)
   parsed.set('pathname', `api/v2/sendtx/${hex}`)
 
-  const headers = {
-    'api-key': config.nowNodesApiKey
-  }
-  const options = { method: 'GET', headers: headers }
+  const response = await blockbookFetch(io, parsed.href)
 
-  const response = await fetch(parsed.href, options)
-
-  const responseJson = await response.json().catch(async err => {
-    const responseText = await response.text()
-    io.logger.error({
-      err,
-      where: 'sendTransaction JSON parse error',
-      responText: responseText
-    })
-    return {
-      error: 'Failed to parse JSON from wrapped blockbook server'
-    }
-  })
-
-  if (responseJson.error != null) {
-    return {
-      id: data.id,
-      error: {
-        message: responseJson.error
-      }
-    }
-  }
-
-  const out: JsonRpcResponse = {
+  return {
     id: data.id,
-    data: responseJson
+    ...response
   }
-  return out
 }
